fix(advanced-selector): style commands target the stored selector, drop selector:edit-style
